store data is being used to render cards
